Fix bug in tests for fairlearn + 1 more test
We merged #298 and #310 shortly after each other but they contained an
incompatibility that broke the fairlearn tests (the code itself was
fine). This PR fixes this incompatibility.

On top, I added the description argument to add_fairlearn_metric_frame,
to be consistent with all the other methods, and also as a test for it.

Finally, 2 small fixes:

- Added type annotation to transpose argument
- Changed order of arguments in docstring to match order in signature