Resolves #292, supersedes #296
As discussed in #296, instead of dealing with the issue what should
happen if the user sets model_card.model_diagram = False after
initializing with True, just remove the attribute from the Card class.
This way, there cannot be the false impression that the diagram can be
removed from the card by setting the attribute to False.

If users really need to remove the model diagram after enabling it at
first, they can just delete the corresponding section (or make it
invisible).

On top of that, a feature was added that passing a str to model_diagram
will use that str as the section name. Previously, always the default
section would be used (for skops template) or no section added (for
custom templates).