Major refactor: better separation of concerns
Code should be much more straightforward now.

Moreover, it is refactored so that the sink function now gets an
iterator of all nodes, not just of a single node. This is better because
printing of nodes can now happen in context. Just as an example, a node
could now be printed differently if it is the first node of its level.