Suggestion how to simplify serialization of funcs
Description

It looks like the state we save for functions contains unnecessary
information (the class name, which is not used) and duplicate
information (the module name appears twice). This change gets rid of
that unnecessary information.

This also allows to remove the need for ufunc_get_state, which can now
be replaced by function_get_state.

Comment

Maybe I'm missing something and there is a reason for the previous
structure, but if there is, there should be a test to check whatever is
covered by that, or at the very least a comment to explain.

Btw. can't remember who wrote this, could be my fault :)