POC: Proposal for how to deal with protocol update
This PR is a POC that is intended as a basis for discussing how we can
deal with updating the skops persistence protocol.

The problem in general occurs if we make a change to the state that is
stored with an object. For instance, we could store an additional field.
When we try to load old state without that key, there would be an error.
Therefore, we have to somehow adjust our loading code to deal with this
possibility.

First of all, we want to avoid having a bunch of conditionals in our
loading code that checks what protocol is used and then does this or
that thing. This approach would become messy and error prone quite
quickly, because we intend to support old protocols for a very long
time.

This proposal here would allow to write a clean implementation of the
new deserialization code without regard for backwards compatibility. The
old code, however, would not be deleted but instead moved to a different
module and registered with its old protocol number. Then, while loading,
we would check the protocol stored in the schema and use the old code if
there is a match for it. If there is no match, we assume we can safely
use the current protocol instead.

This is just a draft and I haven't tested it on a real example. If we
agree that this is the way forward, I would expand on it and test it
properly. But first we should agree that this is the way to go and check
if there are no problems with this approach.

Below are the instructions with how to deal with a change in protocol:

Every time that a backwards incompatible change to the skops format is made
for the first time within a release, the protocol should be bumped to the next
higher number. The old version of the Node, which knows how to deal with the
old state, should be preserved and registered. Let's give an example:

- There is a BC breaking change in FunctionNode.
- Since it's the first BC breaking change in the skops format in this release,
  bump skops.io._protocol.PROTOCOL (this file) from version X to Y.
- Move the old FunctionNode code into 'skops/io/old/_general_vX.py', where 'X'
  is the old protocol.
- Register the _general_vX.FunctionNode in NODE_TYPE_MAPPING inside of
  _persist.py.

Now, if a user loads a FunctionNode state with version X using skops with
version Y, the old code will be used instead of the new one. For all other
node types, if there is no loader for version X, skops will automatically use
version Y instead.