[WIP] Refactor trusted argument usage
See discussion:

https://discord.com/channels/879548962464493619/1047505653603774584/1085550033207836783

Right now, all nodes are always initialized with trusted=False, so they
are not "aware" of what is trusted and what isn't (which means the
argument might as well not exist). This refactor would pass the actual
trusted argument the user provides to the nodes. As a consequence, the
constructed tree is aware of what nodes are considered trusted or not.

Some unit tests are still failing, so this is WIP.