FIX: Bug in random generator persistence (#331)
One of its children was not persisted correctly (pure dict instead of `DictNode`), which made it impossible to check the untrusted types on it. This fix correctly stores the state as a node.

Note that the old implementation does correctly load the `Generator` instance, but it doesn't work with anything but `trusted=True`.

Tests were extended to uncover this bug.

The old persistence scheme was preserved for protocol vesion 0 -- even though it is buggy, it can work under some circumstances.