[skip ci] [WIP] Persist more objs as pure json
Right now, when using skops persistence, only some primitive types are
stored as json: str, numbers, and null. However, in theory we should
also be able to store some lists and dicts. The advantage of that is
that the storage is much more compact, helping with disk size and
readability.

One issue is that we need to check if the lists and dicts can indeed be
stored as pure json objects. To check this, we make a json roundtrip and
verify that the loaded objects still equals the original object.

On top of that, we also need to explicitly check the types. This is
because a numpy float will == a Python float of the same value, so the
json roundtrip might suggest that a dict with numpy floats was restored
successfully. But that's not true, we get a Python float back, which is
not the same. Checking the types explicitly solves the issue.

The presence of numpy types is, however, also a reason why this feature
helps less than I intially thought: Many sklearn objects will store
numpy objects as attributes, so their __dict__ cannot be stored as a
pure json object.