Update to use latest model card changes
Most notably, the new card class stores the specific section types like
PlotSection in its data. This allows us to more directly modify those
sections, e.g. when a user updates a plot file.

Also use the new get_toc method.

Another change was to add a specific task for updating the title of a
plot, instead of having a big task both for title and content with many
if...else. This is cleaner.