Run check for file size difference skops vs pickle
We already measure the runtime performance difference, but we don't
check the file size difference. This PR adds another, similar script
that does exactly that. The checks are run in the same workflow as the
runtime performance checks.

The results are reported, showing top 10 largest differences, once in
terms of absolute, once in terms of relative differences. In contrast to
the runtime performance check, there is never any error raised, no
matter how big the difference is, because it is unclear what, if any,
difference, would count as unacceptable.

For skops, the zip file is highly compressed. This is a reasonable
choice for the benchmark because we can assume that if file size is a
concern, users would choose that option.