Fix bug when visualizing byte nodes
The visualize function failed when trying to visualize models that had
byte attributes. This is because the node's children would contain raw
bytes, which the function doesn't know how to visualize. Therefore,
children of byte and byte array nodes are now skipped (in addition to
the node types already being skipped).

To test this bug, I added a visualization test to the external packages
like LightGBM, which make use of bytes. I'm not sure if some of the
sklearn estimators could also be candidates, but it would surely be
overkill to test visualizing all of them, whereas the overhead is not so
big for the external packages.