FIX attempt at fixing broken GH workflow
The GitHub action for testing the performance of persistence is broken
since the last two weeks. The error is:

> ModuleNotFoundError: No module named 'skops.io'

This seems to be the same issue we already had previously with broken
installs. Nothing relevant has been changed on the workflow yaml, so
there is really no reason why it should suddenly break.

This PR attempts to fix the issue by making the following changes:

- Use Python 3.10 insteaad of 3.11
- Use pip install with -e flag
- Explicitly call python3.10 instead of just python

A PR hook was added to trigger this workflow, it should be removed
before merging.